refactor: extract lifecycle cli module
